refacto: components indexes renamed
