+++ client/src/components/SearchBar/SearchInput/index.tsx
